+++ lib/app/router/router.dart
@@ -7,7 +7,7 @@ import 'package:app/app/service/service.dart';
 import 'package:app/di/di.dart';
 import 'package:app/feature/error/error.dart';
 import 'package:app/feature/root/root.dart';
-import 'package:app/v2/feature/onboarding/onboarding.dart';
+import 'package:app/v2/feature/onboarding/screen/welcome/welcome_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:logging/logging.dart';
@@ -150,7 +150,7 @@ GoRouter getRouter(BuildContext _) {
         pageBuilder: (context, state) => onboardingTransitionPageBuilder(
           context,
           state,
-          const OnboardingPage(),
+          const WelcomeScreen(),
         ),
         routes: [
           chooseNetworkRoute,
