+++ lib/core/error_handler_factory.dart
@@ -1,4 +1,4 @@
-import 'package:app/core/error_handler/standard_error_handler.dart';
+import 'package:app/utils/factories/error_handler/standard_error_handler.dart';
 import 'package:flutter/cupertino.dart';
 
 PrimaryErrorHandler createPrimaryErrorHandler(BuildContext context) {
