--- lib/v2/feature/onboarding/onboarding.dart
@@ -1,2 +1,0 @@
-export 'view/view.dart';
-export 'widgets/sliding_block_chains.dart';
