+++ lib/v2/feature/onboarding/screen/welcome/welcome_screen_model.dart
@@ -1,0 +1,9 @@
+import 'package:app/v2/feature/onboarding/screen/welcome/welcome_screen.dart';
+import 'package:elementary/elementary.dart';
+
+/// [ElementaryModel] for [WelcomeScreen]
+class WelcomeScreenModel extends ElementaryModel {
+  WelcomeScreenModel(
+    ErrorHandler errorHandler,
+  ) : super(errorHandler: errorHandler);
+}
