--- lib/v2/feature/onboarding/view/view.dart
@@ -1,1 +1,0 @@
-export 'onboarding_page.dart';
